[Feature/#18] 전체 멤버 관리 페이지 수정 기능 구현 (#37)
* feat: 멤버 정보 수정 모달 validation 추가

* fix: 멤버 정보 수정 모달 상태 관리 잘못되는 부분 수정

* chore: 전화번호 관련 검증 로직 추가

* design: 멤버 정보 수정 모달 validation text 레이아웃 안 밀리게 css 수정

* design: 자잘한 디자인 수정

* chore: 멤버 수정 모달 버튼 disabled 기능 추가

* fix: 수정 저장 전 모달을 껐다가 다시 켰을 때 이전 정보가 보이지 않도록 수정

* feat: 멤버 정보 수정 모달 학과 검색 기능 추가

* feat: 멤버 정보 수정 api 연결

* chore: 멤버 정보 수정 시 선택된 학과와 학과 검색 텍스트가 다를 시 에러 토스트 띄우도록 예외 처리

* design: 상세 정보 모달 크기 조정

* chore: 변경된 dto에 따른 타입 수정

* fix: cors 해결을 위해서 vite mkcert 추가

* Revert "fix: cors 해결을 위해서 vite mkcert 추가"

This reverts commit a9ca308d01c373cf62ea364c213cfce139fea8c8.

* fix: 멤버 수정 모달에서 기존 학과 텍스트가 뜨지 않는 문제 해결

* chore: 전체 멤버 관리 페이지 목데이터 삭제

* fix: 멤버 정보 수정 모달에서 학과 정보가 없을 시 빈 문자열 띄우도록 수정

* fix: dto 변경에 따른 자잘한 수정

* fix: dto 변경에 따른 자잘한 수정사항 반영

* fix: Dto 변경에 따른 수정 사항 반영

* fix: 수정 시 전화번호 하이픈 제거하도록 수정

* chore: mutation onSuccess 멘트 수정

* chore: 서버에서 보내주는 에러 메시지 띄우도록 변경

* chore: 디스코드 핸들명 -> 디스코드 사용자명으로 네이밍 통일

* chore: 디스코드 핸들명 -> 디스코드 사용자명으로 네이밍 통일

* docs: eslint 규칙 추가

+++ src/apis/allMemberApi.ts
@@ -1,5 +1,5 @@
 import { apiClient } from ".";
-import { AllMemberListDtoType } from "@/types/dtos/member";
+import { AllMemberListDtoType, DepartmentListDtoType } from "@/types/dtos/member";
 
 export const allMemberApi = {
   getAllMemberList: async (
@@ -9,20 +9,43 @@ export const allMemberApi = {
     searchText: string,
   ): Promise<AllMemberListDtoType> => {
     if (searchText) {
-      const searchUrl = `/members?${searchType}=${searchText}&page=${page}&size=${size}`;
+      const searchUrl = `admin/members?${searchType}=${searchText}&page=${page}&size=${size}`;
 
       const response = await apiClient.get(searchUrl);
       return response.data.content;
     }
 
-    const commonUrl = `/members?page=${page}&size=${size}`;
+    const commonUrl = `admin/members?page=${page}&size=${size}`;
 
     const response = await apiClient.get(commonUrl);
     return response.data.content;
   },
 
   deleteMember: async (memberId: number) => {
-    const response = await apiClient.delete(`/members/${memberId}`);
+    const response = await apiClient.delete(`admin/members/${memberId}`);
+    return response.data;
+  },
+
+  searchDepartmentList: async (searchText: string): Promise<DepartmentListDtoType[]> => {
+    const response = await apiClient.get(
+      `common/members/departments/search?department=${searchText}`,
+    );
+    return response.data;
+  },
+
+  editMemberInfo: async (
+    memberId: number,
+    body: {
+      studentId: string;
+      name: string;
+      phone: string;
+      department: string;
+      email: string;
+      discordUsername: string;
+      nickname: string;
+    },
+  ) => {
+    const response = await apiClient.put(`admin/members/${memberId}`, body);
     return response.data;
   },
 };
